requirements.txt
django>=1.9.6
elasticsearch-dsl>=2.1.0,<6.2.0
django-autofixture==0.12.1
Pillow==4.2.1
sklearn
pandas
